update: replace index page
