+++ frontend/src/router.ts
@@ -8,7 +8,6 @@ export type Path =
   | `/hoge`
   | `/pc`
   | `/vr`
-  | `/title`
   | `/webrtc`
 
 export type Params = {
